Update Spinner component with isVisible prop
Update Spinner component with isVisible prop to make show/hide action more generic.
General refactoring.
