+++ takeoutthetrash-ui/src/js/Components/Common/FetchingStateSpinner.js
@@ -1,8 +1,5 @@
 import React, { Fragment } from "react";
 import PropTypes from "prop-types";
-import { connect } from "react-redux";
-import * as citiesSelectors from "../../Selectors/Cities";
-import * as feedbackSelectors from "../../Selectors/Feedback";
 import SpinnerLegend from "./SpinnerLegend";
 
 const Spinner = () => (
@@ -16,9 +13,8 @@ const Spinner = () => (
   </div>
 );
 
-const FetchingStateSpinner = ({ isFetchingCities, isPostingFeedbackForm }) => {
-  if (!(isFetchingCities || isPostingFeedbackForm)) {
-    // introduce an isVisible prop
+const FetchingStateSpinner = ({ isVisible }) => {
+  if (!isVisible) {
     return null;
   }
 
@@ -42,15 +38,7 @@ const FetchingStateSpinner = ({ isFetchingCities, isPostingFeedbackForm }) => {
 };
 
 FetchingStateSpinner.propTypes = {
-  isFetchingCities: PropTypes.bool,
-  isPostingFeedbackForm: PropTypes.bool,
-  postFailureCount: PropTypes.number,
+  isVisible: PropTypes.bool.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  isFetchingCities: citiesSelectors.isFetchingCities(state),
-  isPostingFeedbackForm: feedbackSelectors.isPostingFeedbackForm(state),
-  postFailureCount: feedbackSelectors.getPostFailureCount(state),
-});
-
-export default connect(mapStateToProps)(FetchingStateSpinner);
+export default FetchingStateSpinner;
